requirements.txt
fastapi
uvicorn
requests
reportlabc